fix: ensure type params have unique names
It appears that TS has a bug in its static analysis of static class
members. If an infered type param in a static class member has the same
name as a type param for the class, TS 2.9 will report an erroroneous
use of class type params in a static method.

I'm going to try to get a smaller reproduction, then will open a TS
issue.