feat: array builder should take a singular specifier
Because arrays by default are of only one thing, we should only take a
single `Validator` to specify the type. If we want an array of multiple
types, then we should pass in a `Union`.

An array that cares about the order of the different types is a tuple,
we should consider adding `tuple` as a potential builder.

BREAKING CHANGE: The contract for `v.array` has changed from taking an
array of `Validator`s to taking a single `Validator`.