chore: lock typescript to 2.8
We cannot use typescript 2.9 because the autogenerated .d.ts files use
the new `import('thing.d.ts')` syntax, which is incompatible with
previous TS versions. In favor of making the lib useable for older TS
versions, we should stay at 2.8 until we _need_ to use something from a
higher version.