test: check that empty arrays pass
While debugging an unrelated issue, I thought there may be a bug
with empty arrays being rejected as valid input. Since there was no
unit test covering this case, I was unable to rule this out. This
wasted some debugging time following a red herring. So here is a unit
test to cover this case so that in the future I only need to check the
unit tests to this out as a potential source of bug.